pyproject.toml
[project]
name = "mcp-server-snowflake"
version = "0.1.0"
description = "MCP server for Snowflake"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  {name = "Carlos Serrano"},
  {name = "Jason Summer"},
  {name = "Tyler White"},
]
dependencies = [
    "beautifulsoup4>=4.13.4",
    "mcp[cli]>=1.8.0",
    "pydantic>=2.11.4",
    "requests>=2.32.3",
    "snowflake-connector-python>=3.15.0",
]

[dependency-groups]
dev = [
    "fast-agent-mcp>=0.2.24",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.8",
]

[build-system]
requires = ["hatchling>=1.0.0"]
build-backend = "hatchling.build"

[project.scripts]
mcp-server-snowflake = "mcp_server_snowflake:main"
